server/requirements.txt
# Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
torch==2.1.1
numpy==1.24.3
